refactor: split geo_algorithms primitive_3d intersections
