+++ model/geo_algorithms/src/intersection/mod.rs
@@ -1,8 +1,8 @@
 //! 交点計算アルゴリズム
 //!
-//! Phase C で `geo_primitives` 側の複数形状間交点計算ロジックを
-//! このモジュール配下へ集約した。
-//! pair_base と primitive_* モジュールが現行の正本である。
+//! 各形状組み合わせの交点計算エントリポイントを提供する。
+//! `pair_base` は共通の基礎計算を担い、`primitive_*` と `nurbs_3d` は
+//! 形状種別ごとの公開 API を提供する。
 
 pub mod nurbs_3d;
 pub mod pair_base;
