+++ model/geo_algorithms/src/intersection/primitive_3d.rs
@@ -1,2704 +1,17 @@
 //! 3D Primitive intersection algorithms
-//!
-//! Phase C Step 2: `geo_primitives` から 3D 交差判定ロジックを
-//! `geo_algorithms` 側へ集約するための受け皿。
-//!
-//! orphan rules により trait 実装ではなく形状ペア free-function を提供する。
-//! `geo_primitives` の `BasicIntersection`/`MultipleIntersection` を呼び出す薄い
-//! ラッパー、または `pair_base` に委譲する。
 
-use crate::{
-    Arc3D, Circle3D, ConicalSolid3D, ConicalSurface3D, CylindricalSurface3D, Ellipse3D,
-    EllipsoidalSolid3D, EllipsoidalSurface3D, InfiniteLine3D, IntersectionResult, LineSegment3D,
-    Plane3D, Point3D, Ray3D, SphericalSolid3D, SphericalSurface3D, TorusSolid3D, TorusSurface3D,
-    Triangle3D, TriangleMesh3D,
-};
-use geo_contracts::{
-    Arc3DEndpoint, Arc3DProperties, Circle3DProperties, ConicalSolid3DContainment,
-    ConicalSolid3DProperties, ConicalSurface3DProperties, CylindricalSurface3DProperties,
-    EllipsoidalSolid3DProperties, InfiniteLine3DProperties, Scalar, SphericalSolid3DProperties,
-    SphericalSurface3DProperties, Triangle3DBoundaryAccess,
-};
-
-fn point_intersection_if<T: Scalar>(point: &Point3D<T>, condition: bool) -> Option<Point3D<T>> {
-    if condition {
-        Some(Point3D::new(point.x(), point.y(), point.z()))
-    } else {
-        None
-    }
-}
-
-fn point_matches_either_segment_endpoint<T: Scalar>(
-    point: Point3D<T>,
-    segment: &LineSegment3D<T>,
-    tolerance: T,
-) -> bool {
-    point.distance_to(&segment.start()) <= tolerance
-        || point.distance_to(&segment.end()) <= tolerance
-}
-
-fn conical_solid3d_contains_point_with_tolerance<T: Scalar>(
-    cone: &ConicalSolid3D<T>,
-    point: &Point3D<T>,
-    tolerance: T,
-) -> bool {
-    ConicalSolid3DContainment::contains_point_tolerance(
-        cone,
-        (point.x(), point.y(), point.z()),
-        tolerance,
-    )
-}
-
-fn triangle3d_vertex_points<T: Scalar>(triangle: &Triangle3D<T>) -> [Point3D<T>; 3] {
-    let (ax, ay, az) = Triangle3DBoundaryAccess::vertex_a(triangle);
-    let (bx, by, bz) = Triangle3DBoundaryAccess::vertex_b(triangle);
-    let (cx, cy, cz) = Triangle3DBoundaryAccess::vertex_c(triangle);
-    [
-        Point3D::new(ax, ay, az),
-        Point3D::new(bx, by, bz),
-        Point3D::new(cx, cy, cz),
-    ]
-}
-
-fn spherical_surface_intersection_parameters<T: Scalar>(
-    start: &Point3D<T>,
-    direction: &crate::Vector3D<T>,
-    sphere: &SphericalSurface3D<T>,
-    tolerance: T,
-) -> Option<(T, T)> {
-    let center_tuple = SphericalSurface3DProperties::center(sphere);
-    let center = Point3D::new(center_tuple.0, center_tuple.1, center_tuple.2);
-    let radius = SphericalSurface3DProperties::radius(sphere);
-    let offset = crate::Vector3D::from_points(&center, start);
-
-    let a = direction.dot(direction);
-    if a.abs() <= tolerance {
-        return None;
-    }
-
-    let b = offset.dot(direction) * (T::ONE + T::ONE);
-    let c = offset.dot(&offset) - radius * radius;
-    let discriminant = b * b - (T::ONE + T::ONE + T::ONE + T::ONE) * a * c;
-    if discriminant < T::ZERO {
-        return None;
-    }
-
-    let sqrt_discriminant = discriminant.sqrt();
-    let two_a = (T::ONE + T::ONE) * a;
-    Some((
-        (-b - sqrt_discriminant) / two_a,
-        (-b + sqrt_discriminant) / two_a,
-    ))
-}
-
-// ── Arc3D ─────────────────────────────────────────────────────────────────────
-
-fn arc3d_point3d_intersection_raw<T: Scalar>(
-    arc: &Arc3D<T>,
-    point: &Point3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    point_intersection_if(
-        point,
-        crate::distance::arc3d_point3d_distance(arc, point) <= tolerance
-            && arc.contains_point_angle(Point3D::new(point.x(), point.y(), point.z())),
-    )
-}
-
-pub fn arc3d_point3d_intersection<T: Scalar>(
-    arc: &Arc3D<T>,
-    point: &Point3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        arc3d_point3d_intersection_raw(arc, point, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-fn arc3d_line_segment3d_intersection_raw<T: Scalar>(
-    arc: &Arc3D<T>,
-    segment: &LineSegment3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    let (sx, sy, sz) = <Arc3D<T> as Arc3DEndpoint<T>>::start_point(arc);
-    let (ex, ey, ez) = <Arc3D<T> as Arc3DEndpoint<T>>::end_point(arc);
-    let arc_start = Point3D::new(sx, sy, sz);
-    let arc_end = Point3D::new(ex, ey, ez);
-    let d_seg_s = crate::distance::arc3d_point3d_distance(arc, &segment.start());
-    let d_seg_e = crate::distance::arc3d_point3d_distance(arc, &segment.end());
-    if d_seg_s <= tolerance {
-        return Some(segment.start());
-    }
-    if d_seg_e <= tolerance {
-        return Some(segment.end());
-    }
-    if point_matches_either_segment_endpoint(arc_start, segment, tolerance) {
-        Some(arc_start)
-    } else if point_matches_either_segment_endpoint(arc_end, segment, tolerance) {
-        Some(arc_end)
-    } else {
-        None
-    }
-}
-
-pub fn arc3d_line_segment3d_intersection<T: Scalar>(
-    arc: &Arc3D<T>,
-    segment: &LineSegment3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        arc3d_line_segment3d_intersection_raw(arc, segment, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-fn arc3d_ray3d_intersection_raw<T: Scalar>(
-    arc: &Arc3D<T>,
-    ray: &Ray3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    let origin = ray.origin();
-    let d = crate::distance::arc3d_point3d_distance(arc, &origin);
-    if d <= tolerance {
-        Some(origin)
-    } else {
-        None
-    }
-}
-
-pub fn arc3d_ray3d_intersection<T: Scalar>(
-    arc: &Arc3D<T>,
-    ray: &Ray3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        arc3d_ray3d_intersection_raw(arc, ray, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-fn arc3d_infinite_line3d_intersection_raw<T: Scalar>(
-    arc: &Arc3D<T>,
-    line: &InfiniteLine3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    let (px, py, pz) = InfiniteLine3DProperties::point(line);
-    let pt = Point3D::new(px, py, pz);
-    let d = crate::distance::arc3d_point3d_distance(arc, &pt);
-    if d <= tolerance {
-        Some(pt)
-    } else {
-        None
-    }
-}
-
-pub fn arc3d_infinite_line3d_intersection<T: Scalar>(
-    arc: &Arc3D<T>,
-    line: &InfiniteLine3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        arc3d_infinite_line3d_intersection_raw(arc, line, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-fn arc3d_arc3d_intersection_raw<T: Scalar>(
-    arc_a: &Arc3D<T>,
-    arc_b: &Arc3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    let (s1x, s1y, s1z) = <Arc3D<T> as Arc3DEndpoint<T>>::start_point(arc_a);
-    let (e1x, e1y, e1z) = <Arc3D<T> as Arc3DEndpoint<T>>::end_point(arc_a);
-    let (s2x, s2y, s2z) = <Arc3D<T> as Arc3DEndpoint<T>>::start_point(arc_b);
-    let (e2x, e2y, e2z) = <Arc3D<T> as Arc3DEndpoint<T>>::end_point(arc_b);
-    let pa = Point3D::new(s1x, s1y, s1z);
-    let pb = Point3D::new(e1x, e1y, e1z);
-    let pc = Point3D::new(s2x, s2y, s2z);
-    let pd = Point3D::new(e2x, e2y, e2z);
-    if crate::Vector3D::from_points(&pa, &pc).magnitude() <= tolerance {
-        return Some(pa);
-    }
-    if crate::Vector3D::from_points(&pa, &pd).magnitude() <= tolerance {
-        return Some(pa);
-    }
-    if crate::Vector3D::from_points(&pb, &pc).magnitude() <= tolerance {
-        return Some(pb);
-    }
-    if crate::Vector3D::from_points(&pb, &pd).magnitude() <= tolerance {
-        return Some(pb);
-    }
-    None
-}
-
-pub fn arc3d_arc3d_intersection<T: Scalar>(
-    arc_a: &Arc3D<T>,
-    arc_b: &Arc3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        arc3d_arc3d_intersection_raw(arc_a, arc_b, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-// ── Circle3D ──────────────────────────────────────────────────────────────────
-
-fn circle3d_point3d_intersection_raw<T: Scalar>(
-    circle: &Circle3D<T>,
-    point: &Point3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    point_intersection_if(
-        point,
-        crate::distance::circle3d_point3d_distance(circle, point) <= tolerance,
-    )
-}
-
-pub fn circle3d_point3d_intersection<T: Scalar>(
-    circle: &Circle3D<T>,
-    point: &Point3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        circle3d_point3d_intersection_raw(circle, point, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-fn circle3d_line_segment3d_intersection_raw<T: Scalar>(
-    circle: &Circle3D<T>,
-    segment: &LineSegment3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    let start = segment.start();
-    if crate::distance::circle3d_point3d_distance(circle, &start) <= tolerance {
-        return Some(start);
-    }
-    let end = segment.end();
-    if crate::distance::circle3d_point3d_distance(circle, &end) <= tolerance {
-        return Some(end);
-    }
-    None
-}
-
-pub fn circle3d_line_segment3d_intersection<T: Scalar>(
-    circle: &Circle3D<T>,
-    segment: &LineSegment3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        circle3d_line_segment3d_intersection_raw(circle, segment, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-fn circle3d_ray3d_intersection_raw<T: Scalar>(
-    circle: &Circle3D<T>,
-    ray: &Ray3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    let origin = ray.origin();
-    if crate::distance::circle3d_point3d_distance(circle, &origin) <= tolerance {
-        Some(origin)
-    } else {
-        None
-    }
-}
-
-pub fn circle3d_ray3d_intersection<T: Scalar>(
-    circle: &Circle3D<T>,
-    ray: &Ray3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        circle3d_ray3d_intersection_raw(circle, ray, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-fn circle3d_infinite_line3d_intersection_raw<T: Scalar>(
-    circle: &Circle3D<T>,
-    line: &InfiniteLine3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    let (px, py, pz) = InfiniteLine3DProperties::point(line);
-    let pt = Point3D::new(px, py, pz);
-    if crate::distance::circle3d_point3d_distance(circle, &pt) <= tolerance {
-        Some(pt)
-    } else {
-        None
-    }
-}
-
-pub fn circle3d_infinite_line3d_intersection<T: Scalar>(
-    circle: &Circle3D<T>,
-    line: &InfiniteLine3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        circle3d_infinite_line3d_intersection_raw(circle, line, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-pub fn circle3d_circle3d_intersection<T: Scalar>(
-    _circle_a: &Circle3D<T>,
-    _circle_b: &Circle3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(None, false, tolerance)
-}
-
-// ── CylindricalSurface3D ──────────────────────────────────────────────────────
-
-fn cylindrical_surface3d_point3d_intersection_raw<T: Scalar>(
-    cyl: &CylindricalSurface3D<T>,
-    point: &Point3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    point_intersection_if(
-        point,
-        crate::distance::cylindrical_surface3d_point3d_distance(cyl, point) <= tolerance,
-    )
-}
-
-pub fn cylindrical_surface3d_point3d_intersection<T: Scalar>(
-    cyl: &CylindricalSurface3D<T>,
-    point: &Point3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        cylindrical_surface3d_point3d_intersection_raw(cyl, point, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-fn cylindrical_surface3d_circle3d_intersection_raw<T: Scalar>(
-    cyl: &CylindricalSurface3D<T>,
-    circle: &Circle3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    let (cx, cy, cz) = Circle3DProperties::center(circle);
-    let center = Point3D::new(cx, cy, cz);
-    let dist = crate::distance::cylindrical_surface3d_point3d_distance(cyl, &center);
-    if dist <= Circle3DProperties::radius(circle) + tolerance {
-        Some(center)
-    } else {
-        None
-    }
-}
-
-pub fn cylindrical_surface3d_circle3d_intersection<T: Scalar>(
-    cyl: &CylindricalSurface3D<T>,
-    circle: &Circle3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        cylindrical_surface3d_circle3d_intersection_raw(cyl, circle, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-fn cylindrical_surface3d_line_segment3d_intersection_raw<T: Scalar>(
-    cyl: &CylindricalSurface3D<T>,
-    segment: &LineSegment3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    let s = segment.start();
-    let e = segment.end();
-    if crate::distance::cylindrical_surface3d_point3d_distance(cyl, &s) <= tolerance {
-        Some(s)
-    } else if crate::distance::cylindrical_surface3d_point3d_distance(cyl, &e) <= tolerance {
-        Some(e)
-    } else {
-        None
-    }
-}
-
-pub fn cylindrical_surface3d_line_segment3d_intersection<T: Scalar>(
-    cyl: &CylindricalSurface3D<T>,
-    segment: &LineSegment3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        cylindrical_surface3d_line_segment3d_intersection_raw(cyl, segment, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-fn cylindrical_surface3d_triangle3d_intersection_raw<T: Scalar>(
-    cyl: &CylindricalSurface3D<T>,
-    triangle: &Triangle3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    triangle3d_vertex_points(triangle)
-        .into_iter()
-        .find(|point| {
-            crate::distance::cylindrical_surface3d_point3d_distance(cyl, point) <= tolerance
-        })
-}
-
-pub fn cylindrical_surface3d_triangle3d_intersection<T: Scalar>(
-    cyl: &CylindricalSurface3D<T>,
-    triangle: &Triangle3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        cylindrical_surface3d_triangle3d_intersection_raw(cyl, triangle, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-fn cylindrical_surface3d_plane3d_intersection_raw<T: Scalar>(
-    cyl: &CylindricalSurface3D<T>,
-    plane: &Plane3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    let (cx, cy, cz) = <CylindricalSurface3D<T> as CylindricalSurface3DProperties<T>>::center(cyl);
-    let center = Point3D::new(cx, cy, cz);
-    let radius = <CylindricalSurface3D<T> as CylindricalSurface3DProperties<T>>::radius(cyl);
-    if plane.distance_to_point(center).abs() <= radius + tolerance {
-        Some(Point3D::new(cx, cy, cz))
-    } else {
-        None
-    }
-}
-
-pub fn cylindrical_surface3d_plane3d_intersection<T: Scalar>(
-    cyl: &CylindricalSurface3D<T>,
-    plane: &Plane3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        cylindrical_surface3d_plane3d_intersection_raw(cyl, plane, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-fn cylindrical_surface3d_cylindrical_surface3d_intersection_raw<T: Scalar>(
-    lhs: &CylindricalSurface3D<T>,
-    rhs: &CylindricalSurface3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    let lhs_center_tuple =
-        <CylindricalSurface3D<T> as CylindricalSurface3DProperties<T>>::center(lhs);
-    let rhs_center_tuple =
-        <CylindricalSurface3D<T> as CylindricalSurface3DProperties<T>>::center(rhs);
-    let lhs_center = Point3D::new(lhs_center_tuple.0, lhs_center_tuple.1, lhs_center_tuple.2);
-    let rhs_center = Point3D::new(rhs_center_tuple.0, rhs_center_tuple.1, rhs_center_tuple.2);
-
-    let lhs_dist = crate::distance::cylindrical_surface3d_point3d_distance(lhs, &rhs_center);
-    let rhs_dist = crate::distance::cylindrical_surface3d_point3d_distance(rhs, &lhs_center);
-
-    let lhs_axis_tuple = <CylindricalSurface3D<T> as CylindricalSurface3DProperties<T>>::axis(lhs);
-    let rhs_axis_tuple = <CylindricalSurface3D<T> as CylindricalSurface3DProperties<T>>::axis(rhs);
-    let lhs_axis = crate::Vector3D::new(lhs_axis_tuple.0, lhs_axis_tuple.1, lhs_axis_tuple.2);
-    let rhs_axis = crate::Vector3D::new(rhs_axis_tuple.0, rhs_axis_tuple.1, rhs_axis_tuple.2);
-    let center_delta = crate::Vector3D::from_points(&lhs_center, &rhs_center);
-
-    let angle_tolerance = if tolerance > T::from_f64(1.0e-6) {
-        tolerance
-    } else {
-        T::from_f64(1.0e-6)
-    };
-    let axes_parallel = (lhs_axis.dot(&rhs_axis).abs() - T::ONE).abs() <= angle_tolerance;
-    let axis_distance = center_delta.cross(&lhs_axis).magnitude();
-    let radii_match = (<CylindricalSurface3D<T> as CylindricalSurface3DProperties<T>>::radius(lhs)
-        - <CylindricalSurface3D<T> as CylindricalSurface3DProperties<T>>::radius(rhs))
-    .abs()
-        <= tolerance;
-    let coincident = axes_parallel && axis_distance <= tolerance && radii_match;
-
-    if lhs_dist <= tolerance || rhs_dist <= tolerance || coincident {
-        Some(Point3D::new(
-            (lhs_center.x() + rhs_center.x()) / (T::ONE + T::ONE),
-            (lhs_center.y() + rhs_center.y()) / (T::ONE + T::ONE),
-            (lhs_center.z() + rhs_center.z()) / (T::ONE + T::ONE),
-        ))
-    } else {
-        None
-    }
-}
-
-pub fn cylindrical_surface3d_cylindrical_surface3d_intersection<T: Scalar>(
-    lhs: &CylindricalSurface3D<T>,
-    rhs: &CylindricalSurface3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        cylindrical_surface3d_cylindrical_surface3d_intersection_raw(lhs, rhs, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-// ── Ellipse3D ─────────────────────────────────────────────────────────────────
-
-fn ellipse3d_point3d_intersection_raw<T: Scalar + From<f64>>(
-    ellipse: &Ellipse3D<T>,
-    point: &Point3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    point_intersection_if(
-        point,
-        crate::distance::ellipse3d_point3d_distance(ellipse, point) <= tolerance,
-    )
-}
-
-pub fn ellipse3d_point3d_intersection<T: Scalar + From<f64>>(
-    ellipse: &Ellipse3D<T>,
-    point: &Point3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        ellipse3d_point3d_intersection_raw(ellipse, point, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-pub fn ellipse3d_circle3d_intersections<T: Scalar + From<f64>>(
-    ellipse: &Ellipse3D<T>,
-    circle: &Circle3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    let (cx, cy, cz) = Circle3DProperties::center(circle);
-    let center = Point3D::new(cx, cy, cz);
-    let dist = crate::distance::ellipse3d_point3d_distance(ellipse, &center);
-    let points = if dist <= Circle3DProperties::radius(circle) + tolerance {
-        vec![center]
-    } else {
-        Vec::new()
-    };
-    IntersectionResult::from_option_points(points, false, tolerance)
-}
-
-pub fn ellipse3d_arc3d_intersections<T: Scalar + From<f64>>(
-    ellipse: &Ellipse3D<T>,
-    arc: &Arc3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    let (cx, cy, cz) = Arc3DProperties::center(arc);
-    let center = Point3D::new(cx, cy, cz);
-    let dist = crate::distance::ellipse3d_point3d_distance(ellipse, &center);
-    let points = if dist <= Arc3DProperties::radius(arc) + tolerance {
-        vec![center]
-    } else {
-        Vec::new()
-    };
-    IntersectionResult::from_option_points(points, false, tolerance)
-}
-
-pub fn ellipse3d_line_segment3d_intersections<T: Scalar + From<f64>>(
-    ellipse: &Ellipse3D<T>,
-    segment: &LineSegment3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    let start = segment.start();
-    let end = segment.end();
-    let mut intersections = Vec::new();
-    if crate::distance::ellipse3d_point3d_distance(ellipse, &start) <= tolerance {
-        intersections.push(start);
-    }
-    if crate::distance::ellipse3d_point3d_distance(ellipse, &end) <= tolerance {
-        intersections.push(end);
-    }
-    IntersectionResult::from_option_points(intersections, false, tolerance)
-}
-
-pub fn ellipse3d_infinite_line3d_intersections<T: Scalar + From<f64>>(
-    ellipse: &Ellipse3D<T>,
-    line: &InfiniteLine3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    let (px, py, pz) = InfiniteLine3DProperties::point(line);
-    let point = Point3D::new(px, py, pz);
-    let dist = crate::distance::ellipse3d_point3d_distance(ellipse, &point);
-    let points = if dist <= tolerance {
-        vec![point]
-    } else {
-        Vec::new()
-    };
-    IntersectionResult::from_option_points(points, false, tolerance)
-}
-
-pub fn ellipse3d_ray3d_intersections<T: Scalar + From<f64>>(
-    ellipse: &Ellipse3D<T>,
-    ray: &Ray3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    let o = ray.origin();
-    let dist = crate::distance::ellipse3d_point3d_distance(ellipse, &o);
-    let points = if dist <= tolerance {
-        vec![o]
-    } else {
-        Vec::new()
-    };
-    IntersectionResult::from_option_points(points, false, tolerance)
-}
-
-fn ellipse3d_plane3d_intersection_raw<T: Scalar>(
-    ellipse: &Ellipse3D<T>,
-    plane: &Plane3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    let center = ellipse.center();
-    if plane.distance_to_point(center).abs() <= tolerance {
-        Some(ellipse.center())
-    } else {
-        None
-    }
-}
-
-pub fn ellipse3d_plane3d_intersection<T: Scalar>(
-    ellipse: &Ellipse3D<T>,
-    plane: &Plane3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        ellipse3d_plane3d_intersection_raw(ellipse, plane, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-pub fn ellipse3d_triangle3d_intersections<T: Scalar + From<f64>>(
-    ellipse: &Ellipse3D<T>,
-    triangle: &Triangle3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    let mut intersections = Vec::new();
-    for point in triangle3d_vertex_points(triangle) {
-        if crate::distance::ellipse3d_point3d_distance(ellipse, &point) <= tolerance {
-            intersections.push(point);
-        }
-    }
-    IntersectionResult::from_option_points(intersections, false, tolerance)
-}
-
-pub fn ellipse3d_ellipse3d_intersections<T: Scalar>(
-    _ellipse_a: &Ellipse3D<T>,
-    _ellipse_b: &Ellipse3D<T>,
-    _tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_points(Vec::new(), false, _tolerance)
-}
-
-// ── EllipsoidalSolid3D ────────────────────────────────────────────────────────
-
-fn ellipsoidal_solid3d_point3d_intersection_raw<T: Scalar>(
-    ellipsoid: &EllipsoidalSolid3D<T>,
-    point: &Point3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    point_intersection_if(
-        point,
-        crate::distance::ellipsoidal_solid3d_point3d_distance(ellipsoid, point) <= tolerance,
-    )
-}
-
-pub fn ellipsoidal_solid3d_point3d_intersection<T: Scalar>(
-    ellipsoid: &EllipsoidalSolid3D<T>,
-    point: &Point3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        ellipsoidal_solid3d_point3d_intersection_raw(ellipsoid, point, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-fn ellipsoidal_solid3d_plane3d_intersection_raw<T: Scalar>(
-    ellipsoid: &EllipsoidalSolid3D<T>,
-    plane: &Plane3D<T>,
-    _tolerance: T,
-) -> Option<Point3D<T>> {
-    let (c0, c1, c2) = EllipsoidalSolid3DProperties::center(ellipsoid);
-    let center = Point3D::new(c0, c1, c2);
-    let max_radius = EllipsoidalSolid3DProperties::a_radius(ellipsoid)
-        .max(EllipsoidalSolid3DProperties::b_radius(ellipsoid))
-        .max(EllipsoidalSolid3DProperties::c_radius(ellipsoid));
-    if plane.distance_to_point(center).abs() <= max_radius {
-        Some(Point3D::new(c0, c1, c2))
-    } else {
-        None
-    }
-}
-
-pub fn ellipsoidal_solid3d_plane3d_intersection<T: Scalar>(
-    ellipsoid: &EllipsoidalSolid3D<T>,
-    plane: &Plane3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        ellipsoidal_solid3d_plane3d_intersection_raw(ellipsoid, plane, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-pub fn ellipsoidal_solid3d_infinite_line3d_intersections<T: Scalar>(
-    ellipsoid: &EllipsoidalSolid3D<T>,
-    line: &InfiniteLine3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    let (px, py, pz) = InfiniteLine3DProperties::point(line);
-    let point_on_line = Point3D::new(px, py, pz);
-    let points = if crate::distance::ellipsoidal_solid3d_point3d_distance(ellipsoid, &point_on_line)
-        <= tolerance
-    {
-        vec![point_on_line]
-    } else {
-        vec![]
-    };
-    IntersectionResult::from_option_points(points, false, tolerance)
-}
-
-pub fn ellipsoidal_solid3d_ray3d_intersections<T: Scalar>(
-    ellipsoid: &EllipsoidalSolid3D<T>,
-    ray: &Ray3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    let origin = ray.origin();
-    let points =
-        if crate::distance::ellipsoidal_solid3d_point3d_distance(ellipsoid, &origin) <= tolerance {
-            vec![origin]
-        } else {
-            vec![]
-        };
-    IntersectionResult::from_option_points(points, false, tolerance)
-}
-
-// ── EllipsoidalSurface3D ──────────────────────────────────────────────────────
-
-fn ellipsoidal_surface3d_point3d_intersection_raw<T: Scalar>(
-    ellipsoid: &EllipsoidalSurface3D<T>,
-    point: &Point3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    point_intersection_if(
-        point,
-        crate::distance::ellipsoidal_surface3d_point3d_distance(ellipsoid, point) <= tolerance,
-    )
-}
-
-pub fn ellipsoidal_surface3d_point3d_intersection<T: Scalar>(
-    ellipsoid: &EllipsoidalSurface3D<T>,
-    point: &Point3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        ellipsoidal_surface3d_point3d_intersection_raw(ellipsoid, point, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-// ── ConicalSolid3D ────────────────────────────────────────────────────────────
-
-fn conical_solid3d_point3d_intersection_raw<T: Scalar>(
-    cone: &ConicalSolid3D<T>,
-    point: &Point3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    point_intersection_if(
-        point,
-        conical_solid3d_contains_point_with_tolerance(cone, point, tolerance),
-    )
-}
-
-pub fn conical_solid3d_point3d_intersection<T: Scalar>(
-    cone: &ConicalSolid3D<T>,
-    point: &Point3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        conical_solid3d_point3d_intersection_raw(cone, point, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-fn conical_solid3d_line3d_intersection_raw<T: Scalar>(
-    cone: &ConicalSolid3D<T>,
-    line: &InfiniteLine3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    let (px, py, pz) = InfiniteLine3DProperties::point(line);
-    let point_on_line = Point3D::new(px, py, pz);
-    point_intersection_if(
-        &point_on_line,
-        conical_solid3d_contains_point_with_tolerance(cone, &point_on_line, tolerance),
-    )
-}
-
-pub fn conical_solid3d_line3d_intersection<T: Scalar>(
-    cone: &ConicalSolid3D<T>,
-    line: &InfiniteLine3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        conical_solid3d_line3d_intersection_raw(cone, line, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-fn conical_solid3d_ray3d_intersection_raw<T: Scalar>(
-    cone: &ConicalSolid3D<T>,
-    ray: &Ray3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    let origin = ray.origin();
-    point_intersection_if(
-        &origin,
-        conical_solid3d_contains_point_with_tolerance(cone, &origin, tolerance),
-    )
-}
-
-pub fn conical_solid3d_ray3d_intersection<T: Scalar>(
-    cone: &ConicalSolid3D<T>,
-    ray: &Ray3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        conical_solid3d_ray3d_intersection_raw(cone, ray, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-fn conical_solid3d_line_segment3d_intersection_raw<T: Scalar>(
-    cone: &ConicalSolid3D<T>,
-    segment: &LineSegment3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    let start = segment.start();
-    point_intersection_if(
-        &start,
-        conical_solid3d_contains_point_with_tolerance(cone, &start, tolerance),
-    )
-}
-
-pub fn conical_solid3d_line_segment3d_intersection<T: Scalar>(
-    cone: &ConicalSolid3D<T>,
-    segment: &LineSegment3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        conical_solid3d_line_segment3d_intersection_raw(cone, segment, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-fn conical_solid3d_plane3d_intersection_raw<T: Scalar>(
-    cone: &ConicalSolid3D<T>,
-    plane: &Plane3D<T>,
-    _tolerance: T,
-) -> Option<Point3D<T>> {
-    let (ax, ay, az) = ConicalSolid3DProperties::apex(cone);
-    let (bx, by, bz) = ConicalSolid3DProperties::base_center(cone);
-    let apex = Point3D::new(ax, ay, az);
-    let base = Point3D::new(bx, by, bz);
-    let radius = ConicalSolid3DProperties::radius(cone);
-    let dist_apex = plane.distance_to_point(apex);
-    let dist_base = plane.distance_to_point(base);
-    if dist_apex * dist_base <= T::ZERO || dist_base.abs() <= radius {
-        Some(base)
-    } else {
-        None
-    }
-}
-
-pub fn conical_solid3d_plane3d_intersection<T: Scalar>(
-    cone: &ConicalSolid3D<T>,
-    plane: &Plane3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        conical_solid3d_plane3d_intersection_raw(cone, plane, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-// ── ConicalSurface3D ──────────────────────────────────────────────────────────
-
-/// 円錐面と直線の交差パラメータを解析的に求める（2次方程式）。
-/// 返値: 直線 P(t) = origin + t*direction の有効な t 値 (最大2個)。
-/// `ConicalSurface3DProperties::apex()` は実装上 center (参照点) を返すため、
-/// `height()` を用いて幾何学的頂点を再構築している。
-fn conical_surface3d_intersect_params<T: Scalar>(
-    cone: &ConicalSurface3D<T>,
-    origin: &Point3D<T>,
-    direction: &crate::Vector3D<T>,
-    tolerance: T,
-) -> [Option<T>; 2] {
-    let (cx, cy, cz) = ConicalSurface3DProperties::apex(cone);
-    let h = ConicalSurface3DProperties::height(cone);
-    let (dx, dy, dz) = ConicalSurface3DProperties::axis(cone);
-    let r = ConicalSurface3DProperties::radius(cone);
-
-    if h.abs() <= tolerance || r.abs() <= tolerance {
-        return [None, None];
-    }
-
-    // 幾何学的頂点: center - height * axis
-    let apex_x = cx - h * dx;
-    let apex_y = cy - h * dy;
-    let apex_z = cz - h * dz;
-
-    // cos²(semi_angle) = h² / (h² + r²)
-    let k = (h * h) / (h * h + r * r);
-
-    // δ = origin - apex
-    let det_x = origin.x() - apex_x;
-    let det_y = origin.y() - apex_y;
-    let det_z = origin.z() - apex_z;
-
-    let ld = direction.x() * dx + direction.y() * dy + direction.z() * dz;
-    let dd = det_x * dx + det_y * dy + det_z * dz;
-    let ll = direction.x() * direction.x()
-        + direction.y() * direction.y()
-        + direction.z() * direction.z();
-    let dl = det_x * direction.x() + det_y * direction.y() + det_z * direction.z();
-    let dd_sq = det_x * det_x + det_y * det_y + det_z * det_z;
-
-    let two = T::ONE + T::ONE;
-    let four = two + two;
-    let qa = ld * ld - k * ll;
-    let qb = two * (dd * ld - k * dl);
-    let qc = dd * dd - k * dd_sq;
-
-    if qa.abs() <= tolerance {
-        if qb.abs() > tolerance {
-            return [Some(-qc / qb), None];
-        }
-        return [None, None];
-    }
-
-    let discriminant = qb * qb - four * qa * qc;
-    if discriminant < T::ZERO {
-        return [None, None];
-    }
-
-    let sqrt_d = discriminant.sqrt();
-    let two_a = two * qa;
-    [Some((-qb - sqrt_d) / two_a), Some((-qb + sqrt_d) / two_a)]
-}
-
-fn conical_surface3d_filter_params<T: Scalar>(
-    cone: &ConicalSurface3D<T>,
-    origin: &Point3D<T>,
-    direction: &crate::Vector3D<T>,
-    params: [Option<T>; 2],
-    t_min: T,
-    t_max: T,
-    tolerance: T,
-) -> Vec<Point3D<T>> {
-    let mut pts = Vec::new();
-    for maybe_t in params {
-        let Some(t) = maybe_t else { continue };
-        if t < t_min - tolerance || t > t_max + tolerance {
-            continue;
-        }
-        let p = Point3D::new(
-            origin.x() + t * direction.x(),
-            origin.y() + t * direction.y(),
-            origin.z() + t * direction.z(),
-        );
-        if cone.contains_point(&p, tolerance) {
-            pts.push(p);
-        }
-    }
-    if pts.len() == 2 && pts[0].distance_to(&pts[1]) <= tolerance {
-        pts.truncate(1);
-    }
-    pts
-}
-
-fn conical_surface3d_point3d_intersection_raw<T: Scalar>(
-    cone: &ConicalSurface3D<T>,
-    point: &Point3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    point_intersection_if(
-        point,
-        crate::distance::conical_surface3d_point3d_distance(cone, point) <= tolerance,
-    )
-}
-
-pub fn conical_surface3d_point3d_intersection<T: Scalar>(
-    cone: &ConicalSurface3D<T>,
-    point: &Point3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        conical_surface3d_point3d_intersection_raw(cone, point, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-pub fn conical_surface3d_infinite_line3d_intersections<T: Scalar>(
-    cone: &ConicalSurface3D<T>,
-    line: &InfiniteLine3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    let (px, py, pz) = InfiniteLine3DProperties::point(line);
-    let (lx, ly, lz) = InfiniteLine3DProperties::direction(line);
-    let origin = Point3D::new(px, py, pz);
-    let direction = crate::Vector3D::new(lx, ly, lz);
-    let params = conical_surface3d_intersect_params(cone, &origin, &direction, tolerance);
-    let t_inf = T::ONE / tolerance;
-    IntersectionResult::from_option_points(
-        conical_surface3d_filter_params(
-            cone, &origin, &direction, params, -t_inf, t_inf, tolerance,
-        ),
-        false,
-        tolerance,
-    )
-}
-
-pub fn conical_surface3d_ray3d_intersections<T: Scalar>(
-    cone: &ConicalSurface3D<T>,
-    ray: &Ray3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    let origin = ray.origin();
-    let direction = ray.direction_vector();
-    let params = conical_surface3d_intersect_params(cone, &origin, &direction, tolerance);
-    let t_inf = T::ONE / tolerance;
-    IntersectionResult::from_option_points(
-        conical_surface3d_filter_params(
-            cone,
-            &origin,
-            &direction,
-            params,
-            T::ZERO,
-            t_inf,
-            tolerance,
-        ),
-        false,
-        tolerance,
-    )
-}
-
-pub fn conical_surface3d_line_segment3d_intersections<T: Scalar>(
-    cone: &ConicalSurface3D<T>,
-    segment: &LineSegment3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    let start = segment.start();
-    let end = segment.end();
-    let direction = crate::Vector3D::from_points(&start, &end);
-    let params = conical_surface3d_intersect_params(cone, &start, &direction, tolerance);
-    IntersectionResult::from_option_points(
-        conical_surface3d_filter_params(
-            cone,
-            &start,
-            &direction,
-            params,
-            T::ZERO,
-            T::ONE,
-            tolerance,
-        ),
-        false,
-        tolerance,
-    )
-}
-
-// ── SphericalSolid3D ─────────────────────────────────────────────────────────
-
-fn spherical_solid3d_point3d_intersection_raw<T: Scalar>(
-    sphere: &SphericalSolid3D<T>,
-    point: &Point3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    point_intersection_if(
-        point,
-        crate::distance::spherical_solid3d_point3d_distance(sphere, point) <= tolerance,
-    )
-}
-
-pub fn spherical_solid3d_point3d_intersection<T: Scalar>(
-    sphere: &SphericalSolid3D<T>,
-    point: &Point3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        spherical_solid3d_point3d_intersection_raw(sphere, point, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-fn spherical_solid3d_line3d_intersection_raw<T: Scalar>(
-    sphere: &SphericalSolid3D<T>,
-    line: &InfiniteLine3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    let (cx, cy, cz) = SphericalSolid3DProperties::center(sphere);
-    let center = Point3D::new(cx, cy, cz);
-    let line_point_tuple = InfiniteLine3DProperties::point(line);
-    let line_dir_tuple = InfiniteLine3DProperties::direction(line);
-    let line_point = Point3D::new(line_point_tuple.0, line_point_tuple.1, line_point_tuple.2);
-    let line_dir = crate::Vector3D::new(line_dir_tuple.0, line_dir_tuple.1, line_dir_tuple.2);
-    if sphere.distance_to_infinite_line(&line_point, &line_dir) <= tolerance {
-        Some(line.project_point(&center))
-    } else {
-        None
-    }
-}
-
-pub fn spherical_solid3d_line3d_intersection<T: Scalar>(
-    sphere: &SphericalSolid3D<T>,
-    line: &InfiniteLine3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        spherical_solid3d_line3d_intersection_raw(sphere, line, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-fn spherical_solid3d_ray3d_intersection_raw<T: Scalar>(
-    sphere: &SphericalSolid3D<T>,
-    ray: &Ray3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    let (cx, cy, cz) = SphericalSolid3DProperties::center(sphere);
-    let center = Point3D::new(cx, cy, cz);
-    let ray_direction = ray.direction_vector();
-    if sphere.distance_to_ray(&ray.origin(), &ray_direction) <= tolerance {
-        let parameter = ray.parameter_for_point(&center);
-        let clamped = if parameter < T::ZERO {
-            T::ZERO
-        } else {
-            parameter
-        };
-        Some(ray.point_at_parameter(clamped))
-    } else {
-        None
-    }
-}
-
-pub fn spherical_solid3d_ray3d_intersection<T: Scalar>(
-    sphere: &SphericalSolid3D<T>,
-    ray: &Ray3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        spherical_solid3d_ray3d_intersection_raw(sphere, ray, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-fn spherical_solid3d_line_segment3d_intersection_raw<T: Scalar>(
-    sphere: &SphericalSolid3D<T>,
-    segment: &LineSegment3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    let (cx, cy, cz) = SphericalSolid3DProperties::center(sphere);
-    let center = Point3D::new(cx, cy, cz);
-    if sphere.distance_to_line_segment(&segment.start(), &segment.end()) <= tolerance {
-        let parameter = segment.line().parameter_for_point(&center);
-        let clamped = if parameter < segment.start_param() {
-            segment.start_param()
-        } else if parameter > segment.end_param() {
-            segment.end_param()
-        } else {
-            parameter
-        };
-        Some(segment.line().point_at_parameter(clamped))
-    } else {
-        None
-    }
-}
-
-pub fn spherical_solid3d_line_segment3d_intersection<T: Scalar>(
-    sphere: &SphericalSolid3D<T>,
-    segment: &LineSegment3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        spherical_solid3d_line_segment3d_intersection_raw(sphere, segment, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-fn spherical_solid3d_plane3d_intersection_raw<T: Scalar>(
-    sphere: &SphericalSolid3D<T>,
-    plane: &Plane3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    let (cx, cy, cz) = SphericalSolid3DProperties::center(sphere);
-    let center = Point3D::new(cx, cy, cz);
-    let radius = SphericalSolid3DProperties::radius(sphere);
-    if plane.distance_to_point(center).abs() <= radius + tolerance {
-        Some(plane.project_point(center))
-    } else {
-        None
-    }
-}
-
-pub fn spherical_solid3d_plane3d_intersection<T: Scalar>(
-    sphere: &SphericalSolid3D<T>,
-    plane: &Plane3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        spherical_solid3d_plane3d_intersection_raw(sphere, plane, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-// ── TorusSolid3D ──────────────────────────────────────────────────────────────
-
-fn torus_solid3d_point3d_intersection_raw<T: Scalar>(
-    torus: &TorusSolid3D<T>,
-    point: &Point3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    point_intersection_if(
-        point,
-        crate::distance::torus_solid3d_point3d_distance(torus, point) <= tolerance,
-    )
-}
-
-pub fn torus_solid3d_point3d_intersection<T: Scalar>(
-    torus: &TorusSolid3D<T>,
-    point: &Point3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        torus_solid3d_point3d_intersection_raw(torus, point, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-// ── TorusSurface3D ────────────────────────────────────────────────────────────
-
-fn torus_surface3d_point3d_intersection_raw<T: Scalar>(
-    torus: &TorusSurface3D<T>,
-    point: &Point3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    point_intersection_if(
-        point,
-        crate::distance::torus_surface3d_point3d_distance(torus, point) <= tolerance,
-    )
-}
-
-pub fn torus_surface3d_point3d_intersection<T: Scalar>(
-    torus: &TorusSurface3D<T>,
-    point: &Point3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        torus_surface3d_point3d_intersection_raw(torus, point, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-// ── Triangle3D ────────────────────────────────────────────────────────────────
-
-fn triangle3d_point3d_intersection_raw<T: Scalar>(
-    triangle: &Triangle3D<T>,
-    point: &Point3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    point_intersection_if(
-        point,
-        crate::distance::triangle3d_point3d_distance(triangle, point) <= tolerance,
-    )
-}
-
-pub fn triangle3d_point3d_intersection<T: Scalar>(
-    triangle: &Triangle3D<T>,
-    point: &Point3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        triangle3d_point3d_intersection_raw(triangle, point, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-fn triangle3d_line_segment3d_intersection_raw<T: Scalar>(
-    triangle: &Triangle3D<T>,
-    segment: &LineSegment3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    let start = segment.start();
-    let end = segment.end();
-    let dir = end - start;
-    let length = dir.length();
-    if length <= tolerance {
-        return None;
-    }
-    let ray = Ray3D::new(start, dir)?;
-    let point = triangle3d_ray3d_intersection_raw(triangle, &ray, tolerance)?;
-
-    let t = ray.parameter_for_point(&point);
-    if t >= T::ZERO && t <= length {
-        Some(point)
-    } else {
-        None
-    }
-}
-
-pub fn triangle3d_line_segment3d_intersection<T: Scalar>(
-    triangle: &Triangle3D<T>,
-    segment: &LineSegment3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        triangle3d_line_segment3d_intersection_raw(triangle, segment, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-pub fn line_segment3d_triangle3d_intersection<T: Scalar>(
-    segment: &LineSegment3D<T>,
-    triangle: &Triangle3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    triangle3d_line_segment3d_intersection(triangle, segment, tolerance)
-}
-
-fn triangle3d_ray3d_intersection_raw<T: Scalar>(
-    triangle: &Triangle3D<T>,
-    ray: &Ray3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    let va = triangle.vertex_a();
-    let vb = triangle.vertex_b();
-    let vc = triangle.vertex_c();
-
-    let v0 = Point3D::new(va.0, va.1, va.2);
-    let v1 = Point3D::new(vb.0, vb.1, vb.2);
-    let v2 = Point3D::new(vc.0, vc.1, vc.2);
-
-    let edge1 = crate::Vector3D::from_points(&v0, &v1);
-    let edge2 = crate::Vector3D::from_points(&v0, &v2);
-
-    let ray_dir = ray.direction_vector();
-    let h = ray_dir.cross(&edge2);
-    let a = edge1.dot(&h);
-
-    if a.abs() <= tolerance {
-        return None;
-    }
-
-    let f = T::ONE / a;
-    let s = crate::Vector3D::from_points(&v0, &ray.origin());
-    let u = f * s.dot(&h);
-
-    if u < T::ZERO || u > T::ONE {
-        return None;
-    }
-
-    let q = s.cross(&edge1);
-    let v = f * ray_dir.dot(&q);
-
-    if v < T::ZERO || u + v > T::ONE {
-        return None;
-    }
-
-    let t = f * edge2.dot(&q);
-    if t < T::ZERO {
-        return None;
-    }
-
-    Some(Point3D::new(
-        ray.origin().x() + t * ray_dir.x(),
-        ray.origin().y() + t * ray_dir.y(),
-        ray.origin().z() + t * ray_dir.z(),
-    ))
-}
-
-pub fn triangle3d_ray3d_intersection<T: Scalar>(
-    triangle: &Triangle3D<T>,
-    ray: &Ray3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        triangle3d_ray3d_intersection_raw(triangle, ray, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-pub fn ray3d_triangle3d_intersection<T: Scalar>(
-    ray: &Ray3D<T>,
-    triangle: &Triangle3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    triangle3d_ray3d_intersection(triangle, ray, tolerance)
-}
-
-// ── TriangleMesh3D ────────────────────────────────────────────────────────────
-
-fn triangle_mesh3d_point3d_intersection_raw<T: Scalar>(
-    mesh: &TriangleMesh3D<T>,
-    point: &Point3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    point_intersection_if(
-        point,
-        crate::distance::triangle_mesh3d_point3d_distance(mesh, point) <= tolerance,
-    )
-}
-
-pub fn triangle_mesh3d_point3d_intersection<T: Scalar>(
-    mesh: &TriangleMesh3D<T>,
-    point: &Point3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        triangle_mesh3d_point3d_intersection_raw(mesh, point, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-// ── Plane3D ───────────────────────────────────────────────────────────────────
-
-fn plane3d_point3d_intersection_raw<T: Scalar>(
-    plane: &Plane3D<T>,
-    point: &Point3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    point_intersection_if(point, plane.contains_point(*point, tolerance))
-}
-
-pub fn plane3d_point3d_intersection<T: Scalar>(
-    plane: &Plane3D<T>,
-    point: &Point3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        plane3d_point3d_intersection_raw(plane, point, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-fn plane3d_line_segment3d_intersection_raw<T: Scalar>(
-    plane: &Plane3D<T>,
-    segment: &LineSegment3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    let start = segment.start();
-    let end = segment.end();
-    let direction = crate::Vector3D::from_points(&start, &end);
-
-    if direction.is_zero() {
-        return if plane.contains_point(start, tolerance) {
-            Some(start)
-        } else {
-            None
-        };
-    }
-
-    let normal = plane.normal().as_vector();
-    let denom = direction.dot(&normal);
-
-    if denom.abs() <= tolerance {
-        return if plane.contains_point(start, tolerance) {
-            Some(start)
-        } else {
-            None
-        };
-    }
-
-    let to_plane = crate::Vector3D::from_points(&start, &plane.origin());
-    let t = to_plane.dot(&normal) / denom;
-
-    if t >= T::ZERO && t <= T::ONE {
-        Some(Point3D::new(
-            start.x() + t * direction.x(),
-            start.y() + t * direction.y(),
-            start.z() + t * direction.z(),
-        ))
-    } else {
-        None
-    }
-}
-
-pub fn plane3d_line_segment3d_intersection<T: Scalar>(
-    plane: &Plane3D<T>,
-    segment: &LineSegment3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        plane3d_line_segment3d_intersection_raw(plane, segment, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-fn plane3d_ray3d_intersection_raw<T: Scalar>(
-    plane: &Plane3D<T>,
-    ray: &Ray3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    let origin = ray.origin();
-    let direction = ray.direction_vector();
-    let normal = plane.normal().as_vector();
-    let denom = direction.dot(&normal);
-
-    if denom.abs() <= tolerance {
-        return if plane.contains_point(origin, tolerance) {
-            Some(origin)
-        } else {
-            None
-        };
-    }
-
-    let to_plane = crate::Vector3D::from_points(&origin, &plane.origin());
-    let t = to_plane.dot(&normal) / denom;
-
-    if t >= T::ZERO {
-        Some(Point3D::new(
-            origin.x() + t * direction.x(),
-            origin.y() + t * direction.y(),
-            origin.z() + t * direction.z(),
-        ))
-    } else {
-        None
-    }
-}
-
-pub fn plane3d_ray3d_intersection<T: Scalar>(
-    plane: &Plane3D<T>,
-    ray: &Ray3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        plane3d_ray3d_intersection_raw(plane, ray, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-fn plane3d_infinite_line3d_intersection_raw<T: Scalar>(
-    plane: &Plane3D<T>,
-    line: &InfiniteLine3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    let lp = line.point();
-    let ld = line.direction();
-    let line_point = Point3D::new(lp.0, lp.1, lp.2);
-    let line_dir_vec = crate::Vector3D::new(ld.0, ld.1, ld.2);
-    let normal = plane.normal().as_vector();
-    let denom = line_dir_vec.dot(&normal);
-
-    if denom.abs() <= tolerance {
-        return if plane.distance_to_point(line_point).abs() <= tolerance {
-            Some(line_point)
-        } else {
-            None
-        };
-    }
-
-    let to_plane = crate::Vector3D::from_points(&line_point, &plane.origin());
-    let t = to_plane.dot(&normal) / denom;
-    Some(Point3D::new(
-        line_point.x() + t * line_dir_vec.x(),
-        line_point.y() + t * line_dir_vec.y(),
-        line_point.z() + t * line_dir_vec.z(),
-    ))
-}
-
-pub fn plane3d_infinite_line3d_intersection<T: Scalar>(
-    plane: &Plane3D<T>,
-    line: &InfiniteLine3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        plane3d_infinite_line3d_intersection_raw(plane, line, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-// ── Ray3D ─────────────────────────────────────────────────────────────────────
-
-fn ray3d_point3d_intersection_raw<T: Scalar>(
-    ray: &Ray3D<T>,
-    point: &Point3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    point_intersection_if(point, ray.contains_point(point, tolerance))
-}
-
-pub fn ray3d_point3d_intersection<T: Scalar>(
-    ray: &Ray3D<T>,
-    point: &Point3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        ray3d_point3d_intersection_raw(ray, point, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-pub fn ray3d_spherical_surface3d_intersections<T: Scalar>(
-    ray: &Ray3D<T>,
-    sphere: &SphericalSurface3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    let mut intersections = Vec::new();
-    let start = ray.origin();
-    let direction = ray.direction_vector();
-
-    let Some((t1, t2)) =
-        spherical_surface_intersection_parameters(&start, &direction, sphere, tolerance)
-    else {
-        return IntersectionResult::from_option_points(intersections, false, tolerance);
-    };
-
-    if t1 >= T::ZERO {
-        intersections.push(Point3D::new(
-            start.x() + t1 * direction.x(),
-            start.y() + t1 * direction.y(),
-            start.z() + t1 * direction.z(),
-        ));
-    }
-
-    if t2 >= T::ZERO && (t2 - t1).abs() > tolerance {
-        intersections.push(Point3D::new(
-            start.x() + t2 * direction.x(),
-            start.y() + t2 * direction.y(),
-            start.z() + t2 * direction.z(),
-        ));
-    }
-
-    IntersectionResult::from_option_points(intersections, false, tolerance)
-}
-
-fn ray3d_ray3d_intersection_raw<T: Scalar>(
-    ray_a: &Ray3D<T>,
-    ray_b: &Ray3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    let origin_a = ray_a.origin();
-    let origin_b = ray_b.origin();
-    let direction_a = ray_a.direction_vector();
-    let direction_b = ray_b.direction_vector();
-    let origin_offset = crate::Vector3D::from_points(&origin_b, &origin_a);
-
-    let a = direction_a.dot(&direction_a);
-    let b = direction_a.dot(&direction_b);
-    let c = direction_b.dot(&direction_b);
-    let d = direction_a.dot(&origin_offset);
-    let e = direction_b.dot(&origin_offset);
-
-    let denominator = a * c - b * b;
-    if denominator.abs() <= tolerance {
-        return None;
-    }
-
-    let s = (b * e - c * d) / denominator;
-    let t = (a * e - b * d) / denominator;
-    if s < T::ZERO || t < T::ZERO {
-        return None;
-    }
-
-    let point_a = ray_a.point_at_parameter(s);
-    let point_b = ray_b.point_at_parameter(t);
-    if point_a.distance_to(&point_b) <= tolerance {
-        Some(point_a)
-    } else {
-        None
-    }
-}
-
-pub fn ray3d_ray3d_intersection<T: Scalar>(
-    ray_a: &Ray3D<T>,
-    ray_b: &Ray3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        ray3d_ray3d_intersection_raw(ray_a, ray_b, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-fn ray3d_line_segment3d_intersection_raw<T: Scalar>(
-    ray: &Ray3D<T>,
-    segment: &LineSegment3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    let ray_line = InfiniteLine3D::new(ray.origin(), ray.direction_vector())?;
-    let segment_line = segment.line();
-
-    if ray_line.is_parallel_to(segment_line) || !ray_line.is_coplanar_with(segment_line) {
-        return None;
-    }
-
-    let point = ray_line.intersection_with_line(segment_line)?;
-    if ray.contains_point(&point, tolerance) && segment.contains_point(&point, tolerance) {
-        Some(point)
-    } else {
-        None
-    }
-}
-
-pub fn ray3d_line_segment3d_intersection<T: Scalar>(
-    ray: &Ray3D<T>,
-    segment: &LineSegment3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        ray3d_line_segment3d_intersection_raw(ray, segment, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-fn ray3d_infinite_line3d_intersection_raw<T: Scalar>(
-    ray: &Ray3D<T>,
-    line: &InfiniteLine3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    let ray_line = InfiniteLine3D::new(ray.origin(), ray.direction_vector())?;
-
-    if ray_line.is_parallel_to(line) || !ray_line.is_coplanar_with(line) {
-        return None;
-    }
-
-    let point = ray_line.intersection_with_line(line)?;
-    if ray.contains_point(&point, tolerance) {
-        Some(point)
-    } else {
-        None
-    }
-}
-
-pub fn ray3d_infinite_line3d_intersection<T: Scalar>(
-    ray: &Ray3D<T>,
-    line: &InfiniteLine3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        ray3d_infinite_line3d_intersection_raw(ray, line, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-pub fn ray3d_plane3d_intersection<T: Scalar>(
-    ray: &Ray3D<T>,
-    plane: &Plane3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    plane3d_ray3d_intersection(plane, ray, tolerance)
-}
-
-// ── LineSegment3D ─────────────────────────────────────────────────────────────
-
-fn line_segment3d_point3d_intersection_raw<T: Scalar>(
-    segment: &LineSegment3D<T>,
-    point: &Point3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    point_intersection_if(point, segment.contains_point(point, tolerance))
-}
-
-pub fn line_segment3d_point3d_intersection<T: Scalar>(
-    segment: &LineSegment3D<T>,
-    point: &Point3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        line_segment3d_point3d_intersection_raw(segment, point, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-pub fn line_segment3d_spherical_surface3d_intersections<T: Scalar>(
-    segment: &LineSegment3D<T>,
-    sphere: &SphericalSurface3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    let mut intersections = Vec::new();
-    let start = segment.start();
-    let end = segment.end();
-    let direction = crate::Vector3D::from_points(&start, &end);
-
-    let Some((t1, t2)) =
-        spherical_surface_intersection_parameters(&start, &direction, sphere, tolerance)
-    else {
-        return IntersectionResult::from_option_points(intersections, false, tolerance);
-    };
-
-    if t1 >= T::ZERO && t1 <= T::ONE {
-        intersections.push(Point3D::new(
-            start.x() + t1 * direction.x(),
-            start.y() + t1 * direction.y(),
-            start.z() + t1 * direction.z(),
-        ));
-    }
-
-    if t2 >= T::ZERO && t2 <= T::ONE && (t2 - t1).abs() > tolerance {
-        intersections.push(Point3D::new(
-            start.x() + t2 * direction.x(),
-            start.y() + t2 * direction.y(),
-            start.z() + t2 * direction.z(),
-        ));
-    }
-
-    IntersectionResult::from_option_points(intersections, false, tolerance)
-}
-
-fn line_segment3d_line_segment3d_intersection_raw<T: Scalar>(
-    seg_a: &LineSegment3D<T>,
-    seg_b: &LineSegment3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    let p1 = seg_a.start();
-    let p2 = seg_a.end();
-    let p3 = seg_b.start();
-    let p4 = seg_b.end();
-
-    let d1 = crate::Vector3D::from_points(&p1, &p2);
-    let d2 = crate::Vector3D::from_points(&p3, &p4);
-    let r = crate::Vector3D::from_points(&p3, &p1);
-
-    let a = d1.dot(&d1);
-    let b = d1.dot(&d2);
-    let c = d2.dot(&d2);
-    let d = d1.dot(&r);
-    let e = d2.dot(&r);
-
-    let denom = a * c - b * b;
-    if denom.abs() <= tolerance {
-        return None;
-    }
-
-    let s = (b * e - c * d) / denom;
-    let t = (a * e - b * d) / denom;
-
-    if s >= T::ZERO && s <= T::ONE && t >= T::ZERO && t <= T::ONE {
-        let point1 = Point3D::new(
-            p1.x() + s * d1.x(),
-            p1.y() + s * d1.y(),
-            p1.z() + s * d1.z(),
-        );
-        let point2 = Point3D::new(
-            p3.x() + t * d2.x(),
-            p3.y() + t * d2.y(),
-            p3.z() + t * d2.z(),
-        );
-
-        if point1.distance_to(&point2) <= tolerance {
-            Some(point1)
-        } else {
-            None
-        }
-    } else {
-        None
-    }
-}
-
-pub fn line_segment3d_line_segment3d_intersection<T: Scalar>(
-    seg_a: &LineSegment3D<T>,
-    seg_b: &LineSegment3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        line_segment3d_line_segment3d_intersection_raw(seg_a, seg_b, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-pub fn line_segment3d_ray3d_intersection<T: Scalar>(
-    segment: &LineSegment3D<T>,
-    ray: &Ray3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    ray3d_line_segment3d_intersection(ray, segment, tolerance)
-}
-
-pub fn line_segment3d_infinite_line3d_intersection<T: Scalar>(
-    segment: &LineSegment3D<T>,
-    line: &InfiniteLine3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    infinite_line3d_line_segment3d_intersection(line, segment, tolerance)
-}
-
-pub fn line_segment3d_plane3d_intersection<T: Scalar>(
-    segment: &LineSegment3D<T>,
-    plane: &Plane3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    plane3d_line_segment3d_intersection(plane, segment, tolerance)
-}
-
-// ── InfiniteLine3D ────────────────────────────────────────────────────────────
-
-fn infinite_line3d_point3d_intersection_raw<T: Scalar>(
-    line: &InfiniteLine3D<T>,
-    point: &Point3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    point_intersection_if(point, line.contains_point(point, tolerance))
-}
-
-pub fn infinite_line3d_point3d_intersection<T: Scalar>(
-    line: &InfiniteLine3D<T>,
-    point: &Point3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        infinite_line3d_point3d_intersection_raw(line, point, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-pub fn infinite_line3d_spherical_surface3d_intersections<T: Scalar>(
-    line: &InfiniteLine3D<T>,
-    sphere: &SphericalSurface3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    let mut intersections = Vec::new();
-    let point = line.point();
-    let direction = line.direction();
-    let start = Point3D::new(point.0, point.1, point.2);
-    let direction = crate::Vector3D::new(direction.0, direction.1, direction.2);
-
-    let Some((t1, t2)) =
-        spherical_surface_intersection_parameters(&start, &direction, sphere, tolerance)
-    else {
-        return IntersectionResult::from_option_points(intersections, false, tolerance);
-    };
-
-    intersections.push(Point3D::new(
-        start.x() + t1 * direction.x(),
-        start.y() + t1 * direction.y(),
-        start.z() + t1 * direction.z(),
-    ));
-
-    if (t2 - t1).abs() > tolerance {
-        intersections.push(Point3D::new(
-            start.x() + t2 * direction.x(),
-            start.y() + t2 * direction.y(),
-            start.z() + t2 * direction.z(),
-        ));
-    }
-
-    IntersectionResult::from_option_points(intersections, false, tolerance)
-}
-
-fn infinite_line3d_infinite_line3d_intersection_raw<T: Scalar>(
-    line_a: &InfiniteLine3D<T>,
-    line_b: &InfiniteLine3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    let point = line_a.intersection_with_line(line_b)?;
-    if line_b.distance_to_point(&point) <= tolerance {
-        Some(point)
-    } else {
-        None
-    }
-}
-
-pub fn infinite_line3d_infinite_line3d_intersection<T: Scalar>(
-    line_a: &InfiniteLine3D<T>,
-    line_b: &InfiniteLine3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        infinite_line3d_infinite_line3d_intersection_raw(line_a, line_b, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-fn infinite_line3d_line_segment3d_intersection_raw<T: Scalar>(
-    line: &InfiniteLine3D<T>,
-    segment: &LineSegment3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    let segment_line = segment.line();
-    let point = line.intersection_with_line(segment_line)?;
-    if segment.contains_point(&point, tolerance) {
-        Some(point)
-    } else {
-        None
-    }
-}
-
-pub fn infinite_line3d_line_segment3d_intersection<T: Scalar>(
-    line: &InfiniteLine3D<T>,
-    segment: &LineSegment3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        infinite_line3d_line_segment3d_intersection_raw(line, segment, tolerance),
-        false,
-        tolerance,
-    )
-}
-
-fn infinite_line3d_ray3d_intersection_raw<T: Scalar>(
-    line: &InfiniteLine3D<T>,
-    ray: &Ray3D<T>,
-    tolerance: T,
-) -> Option<Point3D<T>> {
-    let ray_line = InfiniteLine3D::new(ray.origin(), ray.direction_vector())?;
-    let point = line.intersection_with_line(&ray_line)?;
-    if ray.contains_point(&point, tolerance) {
-        Some(point)
-    } else {
-        None
-    }
-}
-
-pub fn infinite_line3d_ray3d_intersection<T: Scalar>(
-    line: &InfiniteLine3D<T>,
-    ray: &Ray3D<T>,
-    tolerance: T,
-) -> IntersectionResult<T> {
-    IntersectionResult::from_option_point(
-        infinite_line3d_ray3d_intersection_raw(line, ray, tolerance),
-        false,
-        tolerance,
-    )
-}
+mod circular_family;
+mod cylindrical_and_conical_family;
+mod linear_family;
+mod planar_and_mesh_family;
+mod shared;
+mod spherical_and_quadric_family;
+
+pub use circular_family::*;
+pub use cylindrical_and_conical_family::*;
+pub use linear_family::*;
+pub use planar_and_mesh_family::*;
+pub use spherical_and_quadric_family::*;
 
 #[cfg(test)]
-mod tests {
-    use super::{
-        arc3d_infinite_line3d_intersection, arc3d_line_segment3d_intersection,
-        arc3d_point3d_intersection, arc3d_ray3d_intersection,
-        circle3d_infinite_line3d_intersection, circle3d_line_segment3d_intersection,
-        circle3d_point3d_intersection, circle3d_ray3d_intersection,
-        cylindrical_surface3d_point3d_intersection, ellipse3d_point3d_intersection,
-        infinite_line3d_line_segment3d_intersection, infinite_line3d_point3d_intersection,
-        infinite_line3d_spherical_surface3d_intersections,
-        line_segment3d_infinite_line3d_intersection, line_segment3d_plane3d_intersection,
-        line_segment3d_point3d_intersection, line_segment3d_ray3d_intersection,
-        line_segment3d_spherical_surface3d_intersections, line_segment3d_triangle3d_intersection,
-        plane3d_line_segment3d_intersection, plane3d_point3d_intersection,
-        plane3d_ray3d_intersection, point_matches_either_segment_endpoint,
-        ray3d_line_segment3d_intersection, ray3d_plane3d_intersection, ray3d_point3d_intersection,
-        ray3d_ray3d_intersection, ray3d_spherical_surface3d_intersections,
-        ray3d_triangle3d_intersection, torus_surface3d_point3d_intersection,
-        triangle3d_line_segment3d_intersection, triangle3d_point3d_intersection,
-        triangle3d_ray3d_intersection, triangle_mesh3d_point3d_intersection,
-    };
-    use crate::{
-        Angle, Arc3D, Circle3D, CylindricalSurface3D, Direction3D, Ellipse3D, InfiniteLine3D,
-        IntersectionGeometry, IntersectionTopology, LineSegment3D, Plane3D, Point3D, Ray3D,
-        SphericalSurface3D, TorusSurface3D, Triangle3D, TriangleMesh3D, Vector3D,
-    };
-    use analysis::test_constants;
-
-    fn standard_distance_tol() -> f64 {
-        test_constants::DISTANCE_TOLERANCE_F64
-    }
-
-    #[test]
-    fn plane_point_intersection_returns_same_point() {
-        let plane = Plane3D::xy_plane(0.0_f64);
-        let point = Point3D::new(1.0, -2.0, 0.0);
-
-        let result = plane3d_point3d_intersection(&plane, &point, standard_distance_tol());
-
-        assert_eq!(result.topology, IntersectionTopology::Crossing);
-        assert!(matches!(result.geometry, IntersectionGeometry::Point(p) if p == point));
-    }
-
-    #[test]
-    fn ray_point_intersection_respects_ray_direction() {
-        let ray = Ray3D::new(Point3D::new(0.0, 0.0, 0.0), Vector3D::new(1.0, 0.0, 0.0)).unwrap();
-        let on_ray = Point3D::new(2.0, 0.0, 0.0);
-        let behind_ray = Point3D::new(-1.0, 0.0, 0.0);
-
-        let on = ray3d_point3d_intersection(&ray, &on_ray, standard_distance_tol());
-        let behind = ray3d_point3d_intersection(&ray, &behind_ray, standard_distance_tol());
-        assert_eq!(on.topology, IntersectionTopology::Crossing);
-        assert!(matches!(on.geometry, IntersectionGeometry::Point(p) if p == on_ray));
-        assert_eq!(behind.topology, IntersectionTopology::Disjoint);
-    }
-
-    #[test]
-    fn line_segment_point_intersection_checks_segment_bounds() {
-        let segment =
-            LineSegment3D::new(Point3D::new(0.0, 0.0, 0.0), Point3D::new(2.0, 0.0, 0.0)).unwrap();
-        let on_segment = Point3D::new(1.0, 0.0, 0.0);
-        let outside_segment = Point3D::new(3.0, 0.0, 0.0);
-
-        let on =
-            line_segment3d_point3d_intersection(&segment, &on_segment, standard_distance_tol());
-        let outside = line_segment3d_point3d_intersection(
-            &segment,
-            &outside_segment,
-            standard_distance_tol(),
-        );
-        assert_eq!(on.topology, IntersectionTopology::Crossing);
-        assert!(matches!(on.geometry, IntersectionGeometry::Point(p) if p == on_segment));
-        assert_eq!(outside.topology, IntersectionTopology::Disjoint);
-    }
-
-    #[test]
-    fn infinite_line_point_intersection_checks_collinearity() {
-        let line = InfiniteLine3D::from_two_points(
-            Point3D::new(0.0, 0.0, 0.0),
-            Point3D::new(1.0, 0.0, 0.0),
-        )
-        .unwrap();
-        let on_line = Point3D::new(5.0, 0.0, 0.0);
-        let off_line = Point3D::new(0.0, 1.0, 0.0);
-
-        let on = infinite_line3d_point3d_intersection(&line, &on_line, standard_distance_tol());
-        let off = infinite_line3d_point3d_intersection(&line, &off_line, standard_distance_tol());
-        assert_eq!(on.topology, IntersectionTopology::Crossing);
-        assert!(matches!(on.geometry, IntersectionGeometry::Point(p) if p == on_line));
-        assert_eq!(off.topology, IntersectionTopology::Disjoint);
-    }
-
-    #[test]
-    fn triangle_point_intersection_uses_distance_based_test() {
-        let triangle = Triangle3D::new(
-            Point3D::new(0.0, 0.0, 0.0),
-            Point3D::new(1.0, 0.0, 0.0),
-            Point3D::new(0.0, 1.0, 0.0),
-        )
-        .unwrap();
-        let on_triangle = Point3D::new(0.2, 0.2, 0.0);
-        let off_triangle = Point3D::new(0.2, 0.2, 0.5);
-
-        let on = triangle3d_point3d_intersection(&triangle, &on_triangle, standard_distance_tol());
-        let off =
-            triangle3d_point3d_intersection(&triangle, &off_triangle, standard_distance_tol());
-        assert_eq!(on.topology, IntersectionTopology::Crossing);
-        assert!(matches!(on.geometry, IntersectionGeometry::Point(p) if p == on_triangle));
-        assert_eq!(off.topology, IntersectionTopology::Disjoint);
-    }
-
-    #[test]
-    fn circle_point_intersection_requires_point_on_circumference() {
-        let circle = Circle3D::new(
-            Point3D::new(0.0, 0.0, 0.0),
-            Direction3D::new(0.0, 0.0, 1.0).unwrap(),
-            2.0,
-        )
-        .unwrap();
-        let on_circle = Point3D::new(2.0, 0.0, 0.0);
-        let inside_disk = Point3D::new(1.0, 0.0, 0.0);
-
-        assert_eq!(
-            circle3d_point3d_intersection(&circle, &on_circle, standard_distance_tol()).topology,
-            IntersectionTopology::Crossing
-        );
-        assert!(matches!(
-            circle3d_point3d_intersection(&circle, &on_circle, standard_distance_tol()).geometry,
-            IntersectionGeometry::Point(point) if point == on_circle
-        ));
-        assert_eq!(
-            circle3d_point3d_intersection(&circle, &inside_disk, standard_distance_tol()).topology,
-            IntersectionTopology::Disjoint
-        );
-        assert!(matches!(
-            circle3d_point3d_intersection(&circle, &inside_disk, standard_distance_tol()).geometry,
-            IntersectionGeometry::None
-        ));
-    }
-
-    #[test]
-    fn circle_point_intersection_result_converts_to_topology() {
-        let circle = Circle3D::new(
-            Point3D::new(0.0, 0.0, 0.0),
-            Direction3D::new(0.0, 0.0, 1.0).unwrap(),
-            2.0,
-        )
-        .unwrap();
-        let on_circle = Point3D::new(2.0, 0.0, 0.0);
-        let off_circle = Point3D::new(1.0, 0.0, 0.0);
-
-        let hit = circle3d_point3d_intersection(&circle, &on_circle, standard_distance_tol());
-        let miss = circle3d_point3d_intersection(&circle, &off_circle, standard_distance_tol());
-
-        assert_eq!(hit.topology, IntersectionTopology::Crossing);
-        assert!(matches!(hit.geometry, IntersectionGeometry::Point(_)));
-        assert_eq!(miss.topology, IntersectionTopology::Disjoint);
-        assert!(matches!(miss.geometry, IntersectionGeometry::None));
-    }
-
-    #[test]
-    fn arc_point_intersection_checks_angle_range() {
-        let arc = Arc3D::new(
-            Point3D::new(0.0, 0.0, 0.0),
-            2.0,
-            Direction3D::new(0.0, 0.0, 1.0).unwrap(),
-            Direction3D::new(1.0, 0.0, 0.0).unwrap(),
-            Angle::from_radians(0.0),
-            Angle::from_radians(std::f64::consts::FRAC_PI_2),
-        )
-        .unwrap();
-
-        let on_arc = Point3D::new(2.0, 0.0, 0.0);
-        let out_of_angle = Point3D::new(-2.0, 0.0, 0.0);
-
-        assert_eq!(
-            arc3d_point3d_intersection(&arc, &on_arc, standard_distance_tol()).topology,
-            IntersectionTopology::Crossing
-        );
-        assert!(matches!(
-            arc3d_point3d_intersection(&arc, &on_arc, standard_distance_tol()).geometry,
-            IntersectionGeometry::Point(point) if point == on_arc
-        ));
-        assert_eq!(
-            arc3d_point3d_intersection(&arc, &out_of_angle, standard_distance_tol()).topology,
-            IntersectionTopology::Disjoint
-        );
-        assert!(matches!(
-            arc3d_point3d_intersection(&arc, &out_of_angle, standard_distance_tol()).geometry,
-            IntersectionGeometry::None
-        ));
-    }
-
-    #[test]
-    fn arc_point_intersection_result_converts_to_topology() {
-        let arc = Arc3D::new(
-            Point3D::new(0.0, 0.0, 0.0),
-            2.0,
-            Direction3D::new(0.0, 0.0, 1.0).unwrap(),
-            Direction3D::new(1.0, 0.0, 0.0).unwrap(),
-            Angle::from_radians(0.0),
-            Angle::from_radians(std::f64::consts::FRAC_PI_2),
-        )
-        .unwrap();
-
-        let on_arc = Point3D::new(2.0, 0.0, 0.0);
-        let off_arc = Point3D::new(-2.0, 0.0, 0.0);
-        let hit = arc3d_point3d_intersection(&arc, &on_arc, standard_distance_tol());
-        let miss = arc3d_point3d_intersection(&arc, &off_arc, standard_distance_tol());
-
-        assert_eq!(hit.topology, IntersectionTopology::Crossing);
-        assert!(matches!(hit.geometry, IntersectionGeometry::Point(_)));
-        assert_eq!(miss.topology, IntersectionTopology::Disjoint);
-        assert!(matches!(miss.geometry, IntersectionGeometry::None));
-    }
-
-    #[test]
-    fn arc_linear_input_result_variants_convert_to_topology() {
-        let arc = Arc3D::new(
-            Point3D::new(0.0, 0.0, 0.0),
-            2.0,
-            Direction3D::new(0.0, 0.0, 1.0).unwrap(),
-            Direction3D::new(1.0, 0.0, 0.0).unwrap(),
-            Angle::from_radians(0.0),
-            Angle::from_radians(std::f64::consts::FRAC_PI_2),
-        )
-        .unwrap();
-
-        let segment_hit =
-            LineSegment3D::new(Point3D::new(2.0, 0.0, 0.0), Point3D::new(3.0, 0.0, 0.0)).unwrap();
-        let segment_miss =
-            LineSegment3D::new(Point3D::new(5.0, 0.0, 0.0), Point3D::new(6.0, 0.0, 0.0)).unwrap();
-        let ray_hit =
-            Ray3D::new(Point3D::new(2.0, 0.0, 0.0), Vector3D::new(1.0, 0.0, 0.0)).unwrap();
-        let ray_miss =
-            Ray3D::new(Point3D::new(5.0, 0.0, 0.0), Vector3D::new(1.0, 0.0, 0.0)).unwrap();
-        let line_hit = InfiniteLine3D::from_two_points(
-            Point3D::new(2.0, 0.0, 0.0),
-            Point3D::new(2.0, 1.0, 0.0),
-        )
-        .unwrap();
-        let line_miss = InfiniteLine3D::from_two_points(
-            Point3D::new(5.0, 0.0, 0.0),
-            Point3D::new(5.0, 1.0, 0.0),
-        )
-        .unwrap();
-
-        let hit_segment =
-            arc3d_line_segment3d_intersection(&arc, &segment_hit, standard_distance_tol());
-        let miss_segment =
-            arc3d_line_segment3d_intersection(&arc, &segment_miss, standard_distance_tol());
-        let hit_ray = arc3d_ray3d_intersection(&arc, &ray_hit, standard_distance_tol());
-        let miss_ray = arc3d_ray3d_intersection(&arc, &ray_miss, standard_distance_tol());
-        let hit_line = arc3d_infinite_line3d_intersection(&arc, &line_hit, standard_distance_tol());
-        let miss_line =
-            arc3d_infinite_line3d_intersection(&arc, &line_miss, standard_distance_tol());
-
-        assert_eq!(hit_segment.topology, IntersectionTopology::Crossing);
-        assert!(matches!(
-            hit_segment.geometry,
-            IntersectionGeometry::Point(_)
-        ));
-        assert_eq!(miss_segment.topology, IntersectionTopology::Disjoint);
-        assert_eq!(hit_ray.topology, IntersectionTopology::Crossing);
-        assert_eq!(miss_ray.topology, IntersectionTopology::Disjoint);
-        assert_eq!(hit_line.topology, IntersectionTopology::Crossing);
-        assert_eq!(miss_line.topology, IntersectionTopology::Disjoint);
-    }
-
-    #[test]
-    fn segment_endpoint_match_helper_is_symmetric() {
-        let segment =
-            LineSegment3D::new(Point3D::new(2.0, 0.0, 0.0), Point3D::new(3.0, 0.0, 0.0)).unwrap();
-
-        assert!(point_matches_either_segment_endpoint(
-            Point3D::new(2.0, 0.0, 0.0),
-            &segment,
-            standard_distance_tol(),
-        ));
-        assert!(point_matches_either_segment_endpoint(
-            Point3D::new(3.0, 0.0, 0.0),
-            &segment,
-            standard_distance_tol(),
-        ));
-        assert!(!point_matches_either_segment_endpoint(
-            Point3D::new(4.0, 0.0, 0.0),
-            &segment,
-            standard_distance_tol(),
-        ));
-    }
-
-    #[test]
-    fn ellipse_point_intersection_uses_distance() {
-        let ellipse = Ellipse3D::new(
-            Point3D::new(0.0, 0.0, 0.0),
-            3.0,
-            2.0,
-            Vector3D::new(0.0, 0.0, 1.0),
-            Vector3D::new(1.0, 0.0, 0.0),
-        )
-        .unwrap();
-
-        let on_ellipse = Point3D::new(3.0, 0.0, 0.0);
-        let inside_ellipse = Point3D::new(1.0, 0.0, 0.0);
-        let outside_plane = Point3D::new(0.0, 0.0, 0.5);
-
-        let on = ellipse3d_point3d_intersection(&ellipse, &on_ellipse, standard_distance_tol());
-        let inside =
-            ellipse3d_point3d_intersection(&ellipse, &inside_ellipse, standard_distance_tol());
-        let outside =
-            ellipse3d_point3d_intersection(&ellipse, &outside_plane, standard_distance_tol());
-        assert_eq!(on.topology, IntersectionTopology::Crossing);
-        assert!(matches!(on.geometry, IntersectionGeometry::Point(p) if p == on_ellipse));
-        assert_eq!(inside.topology, IntersectionTopology::Crossing);
-        assert!(matches!(inside.geometry, IntersectionGeometry::Point(p) if p == inside_ellipse));
-        assert_eq!(outside.topology, IntersectionTopology::Disjoint);
-    }
-
-    #[test]
-    fn circle_linear_input_result_variants_convert_to_topology() {
-        let circle = Circle3D::new(
-            Point3D::new(0.0, 0.0, 0.0),
-            Direction3D::new(0.0, 0.0, 1.0).unwrap(),
-            2.0,
-        )
-        .unwrap();
-
-        let segment_hit =
-            LineSegment3D::new(Point3D::new(2.0, 0.0, 0.0), Point3D::new(3.0, 0.0, 0.0)).unwrap();
-        let segment_miss =
-            LineSegment3D::new(Point3D::new(5.0, 0.0, 0.0), Point3D::new(6.0, 0.0, 0.0)).unwrap();
-        let ray_hit =
-            Ray3D::new(Point3D::new(2.0, 0.0, 0.0), Vector3D::new(1.0, 0.0, 0.0)).unwrap();
-        let ray_miss =
-            Ray3D::new(Point3D::new(5.0, 0.0, 0.0), Vector3D::new(1.0, 0.0, 0.0)).unwrap();
-        let line_hit = InfiniteLine3D::from_two_points(
-            Point3D::new(2.0, 0.0, 0.0),
-            Point3D::new(2.0, 1.0, 0.0),
-        )
-        .unwrap();
-        let line_miss = InfiniteLine3D::from_two_points(
-            Point3D::new(5.0, 0.0, 0.0),
-            Point3D::new(5.0, 1.0, 0.0),
-        )
-        .unwrap();
-
-        let hit_segment =
-            circle3d_line_segment3d_intersection(&circle, &segment_hit, standard_distance_tol());
-        let miss_segment =
-            circle3d_line_segment3d_intersection(&circle, &segment_miss, standard_distance_tol());
-        let hit_ray = circle3d_ray3d_intersection(&circle, &ray_hit, standard_distance_tol());
-        let miss_ray = circle3d_ray3d_intersection(&circle, &ray_miss, standard_distance_tol());
-        let hit_line =
-            circle3d_infinite_line3d_intersection(&circle, &line_hit, standard_distance_tol());
-        let miss_line =
-            circle3d_infinite_line3d_intersection(&circle, &line_miss, standard_distance_tol());
-
-        assert_eq!(hit_segment.topology, IntersectionTopology::Crossing);
-        assert!(matches!(
-            hit_segment.geometry,
-            IntersectionGeometry::Point(_)
-        ));
-        assert_eq!(miss_segment.topology, IntersectionTopology::Disjoint);
-        assert_eq!(hit_ray.topology, IntersectionTopology::Crossing);
-        assert_eq!(miss_ray.topology, IntersectionTopology::Disjoint);
-        assert_eq!(hit_line.topology, IntersectionTopology::Crossing);
-        assert_eq!(miss_line.topology, IntersectionTopology::Disjoint);
-    }
-
-    #[test]
-    fn torus_surface_point_intersection_uses_surface_distance() {
-        let torus = TorusSurface3D::new(
-            Point3D::new(0.0, 0.0, 0.0),
-            Direction3D::new(0.0, 0.0, 1.0).unwrap(),
-            Direction3D::new(1.0, 0.0, 0.0).unwrap(),
-            3.0,
-            1.0,
-        )
-        .unwrap();
-
-        let on_surface = Point3D::new(4.0, 0.0, 0.0);
-        let inside_tube = Point3D::new(3.0, 0.0, 0.0);
-
-        let on = torus_surface3d_point3d_intersection(&torus, &on_surface, standard_distance_tol());
-        let inside =
-            torus_surface3d_point3d_intersection(&torus, &inside_tube, standard_distance_tol());
-        assert_eq!(on.topology, IntersectionTopology::Crossing);
-        assert!(matches!(on.geometry, IntersectionGeometry::Point(p) if p == on_surface));
-        assert_eq!(inside.topology, IntersectionTopology::Disjoint);
-    }
-
-    #[test]
-    fn cylindrical_surface_point_intersection_uses_surface_distance() {
-        let cyl = CylindricalSurface3D::new_z_axis(Point3D::new(0.0, 0.0, 0.0), 1.0).unwrap();
-
-        let on_surface = Point3D::new(1.0, 0.0, 0.5);
-        let off_surface = Point3D::new(2.0, 0.0, 0.5);
-
-        let on =
-            cylindrical_surface3d_point3d_intersection(&cyl, &on_surface, standard_distance_tol());
-        let off =
-            cylindrical_surface3d_point3d_intersection(&cyl, &off_surface, standard_distance_tol());
-        assert_eq!(on.topology, IntersectionTopology::Crossing);
-        assert!(matches!(on.geometry, IntersectionGeometry::Point(p) if p == on_surface));
-        assert_eq!(off.topology, IntersectionTopology::Disjoint);
-    }
-
-    #[test]
-    fn triangle_mesh_point_intersection_checks_member_triangles() {
-        let mesh = TriangleMesh3D::new(
-            vec![
-                Point3D::new(0.0, 0.0, 0.0),
-                Point3D::new(1.0, 0.0, 0.0),
-                Point3D::new(0.0, 1.0, 0.0),
-            ],
-            vec![[0, 1, 2]],
-        )
-        .unwrap();
-
-        let on_triangle = Point3D::new(0.2, 0.2, 0.0);
-        let off_triangle = Point3D::new(0.2, 0.2, 0.3);
-
-        let on = triangle_mesh3d_point3d_intersection(&mesh, &on_triangle, standard_distance_tol());
-        let off =
-            triangle_mesh3d_point3d_intersection(&mesh, &off_triangle, standard_distance_tol());
-        assert_eq!(on.topology, IntersectionTopology::Crossing);
-        assert!(matches!(on.geometry, IntersectionGeometry::Point(p) if p == on_triangle));
-        assert_eq!(off.topology, IntersectionTopology::Disjoint);
-    }
-
-    #[test]
-    fn symmetric_triangle_intersection_wrappers_match_base_functions() {
-        let tri = Triangle3D::new(
-            Point3D::new(0.0, 0.0, 0.0),
-            Point3D::new(1.0, 0.0, 0.0),
-            Point3D::new(0.0, 1.0, 0.0),
-        )
-        .unwrap();
-        let seg =
-            LineSegment3D::new(Point3D::new(0.2, 0.2, -1.0), Point3D::new(0.2, 0.2, 1.0)).unwrap();
-        let ray = Ray3D::new(Point3D::new(0.2, 0.2, 1.0), Vector3D::new(0.0, 0.0, -1.0)).unwrap();
-
-        let tol = standard_distance_tol();
-        assert_eq!(
-            line_segment3d_triangle3d_intersection(&seg, &tri, tol).topology,
-            triangle3d_line_segment3d_intersection(&tri, &seg, tol).topology
-        );
-        assert_eq!(
-            ray3d_triangle3d_intersection(&ray, &tri, tol).topology,
-            triangle3d_ray3d_intersection(&tri, &ray, tol).topology
-        );
-    }
-
-    #[test]
-    fn symmetric_plane_intersection_wrappers_match_base_functions() {
-        let tol = standard_distance_tol();
-        let plane = Plane3D::xy_plane(0.0_f64);
-        let ray = Ray3D::new(Point3D::new(0.0, 0.0, 1.0), Vector3D::new(0.0, 0.0, -1.0)).unwrap();
-        let seg =
-            LineSegment3D::new(Point3D::new(0.0, 0.0, -1.0), Point3D::new(0.0, 0.0, 1.0)).unwrap();
-
-        assert_eq!(
-            ray3d_plane3d_intersection(&ray, &plane, tol).topology,
-            plane3d_ray3d_intersection(&plane, &ray, tol).topology
-        );
-        assert_eq!(
-            line_segment3d_plane3d_intersection(&seg, &plane, tol).topology,
-            plane3d_line_segment3d_intersection(&plane, &seg, tol).topology
-        );
-    }
-
-    #[test]
-    fn symmetric_line_segment_intersection_wrappers_match_base_functions() {
-        let tol = standard_distance_tol();
-        let seg =
-            LineSegment3D::new(Point3D::new(0.0, 0.0, 0.0), Point3D::new(1.0, 0.0, 0.0)).unwrap();
-        let ray = Ray3D::new(Point3D::new(0.0, 0.0, 0.0), Vector3D::new(1.0, 0.0, 0.0)).unwrap();
-        let line = InfiniteLine3D::from_two_points(
-            Point3D::new(0.0, 0.0, 0.0),
-            Point3D::new(1.0, 0.0, 0.0),
-        )
-        .unwrap();
-
-        assert_eq!(
-            line_segment3d_ray3d_intersection(&seg, &ray, tol).topology,
-            ray3d_line_segment3d_intersection(&ray, &seg, tol).topology
-        );
-        assert_eq!(
-            line_segment3d_infinite_line3d_intersection(&seg, &line, tol).topology,
-            infinite_line3d_line_segment3d_intersection(&line, &seg, tol).topology
-        );
-    }
-
-    #[test]
-    fn spherical_surface_line_like_intersections_return_expected_points() {
-        let tolerance = 1e-9;
-        let sphere = SphericalSurface3D::new(
-            Point3D::new(0.0, 0.0, 0.0),
-            Vector3D::new(0.0, 0.0, 1.0),
-            Vector3D::new(1.0, 0.0, 0.0),
-            1.0,
-        )
-        .unwrap();
-
-        let segment =
-            LineSegment3D::new(Point3D::new(-2.0, 0.0, 0.0), Point3D::new(2.0, 0.0, 0.0)).unwrap();
-        let ray = Ray3D::new(Point3D::new(-2.0, 0.0, 0.0), Vector3D::new(1.0, 0.0, 0.0)).unwrap();
-        let line = InfiniteLine3D::from_two_points(
-            Point3D::new(-2.0, 0.0, 0.0),
-            Point3D::new(2.0, 0.0, 0.0),
-        )
-        .unwrap();
-
-        let seg_result =
-            line_segment3d_spherical_surface3d_intersections(&segment, &sphere, tolerance);
-        assert_eq!(seg_result.topology, IntersectionTopology::Crossing);
-        if let IntersectionGeometry::Points(ref pts) = seg_result.geometry {
-            assert_eq!(pts.len(), 2);
-            assert_eq!(pts[0], Point3D::new(-1.0, 0.0, 0.0));
-            assert_eq!(pts[1], Point3D::new(1.0, 0.0, 0.0));
-        } else {
-            panic!("Expected Points geometry");
-        }
-
-        let ray_result = ray3d_spherical_surface3d_intersections(&ray, &sphere, tolerance);
-        assert_eq!(ray_result.topology, IntersectionTopology::Crossing);
-        if let IntersectionGeometry::Points(ref pts) = ray_result.geometry {
-            assert_eq!(pts.len(), 2);
-            assert_eq!(pts[0], Point3D::new(-1.0, 0.0, 0.0));
-            assert_eq!(pts[1], Point3D::new(1.0, 0.0, 0.0));
-        } else {
-            panic!("Expected Points geometry");
-        }
-
-        let line_result =
-            infinite_line3d_spherical_surface3d_intersections(&line, &sphere, tolerance);
-        assert_eq!(line_result.topology, IntersectionTopology::Crossing);
-        if let IntersectionGeometry::Points(ref pts) = line_result.geometry {
-            assert_eq!(pts.len(), 2);
-            assert_eq!(pts[0], Point3D::new(-1.0, 0.0, 0.0));
-            assert_eq!(pts[1], Point3D::new(1.0, 0.0, 0.0));
-        } else {
-            panic!("Expected Points geometry");
-        }
-    }
-
-    #[test]
-    fn ray_ray_intersection_returns_shared_point_only_for_forward_rays() {
-        let tolerance = 1e-9;
-        let ray_a = Ray3D::new(Point3D::new(-1.0, 0.0, 0.0), Vector3D::new(1.0, 0.0, 0.0)).unwrap();
-        let ray_b = Ray3D::new(Point3D::new(0.0, -1.0, 0.0), Vector3D::new(0.0, 1.0, 0.0)).unwrap();
-        let opposite =
-            Ray3D::new(Point3D::new(0.0, -1.0, 0.0), Vector3D::new(0.0, -1.0, 0.0)).unwrap();
-
-        let crossing = ray3d_ray3d_intersection(&ray_a, &ray_b, tolerance);
-        let disjoint = ray3d_ray3d_intersection(&ray_a, &opposite, tolerance);
-        assert_eq!(crossing.topology, IntersectionTopology::Crossing);
-        let origin = Point3D::new(0.0_f64, 0.0, 0.0);
-        if let IntersectionGeometry::Point(ref p) = crossing.geometry {
-            assert!(p.distance_to(&origin) < 1e-9);
-        } else {
-            panic!("Expected IntersectionGeometry::Point for crossing");
-        }
-        assert_eq!(disjoint.topology, IntersectionTopology::Disjoint);
-    }
-
-    #[test]
-    fn conical_solid_point_boundary_guard_keeps_intersection_on_point_helper() {
-        const CONICAL_SOLID_DIRECT_CONTAINMENT: &str =
-            "ConicalSolid3DContainment::contains_point_tolerance";
-        const CONICAL_SOLID_POINT_HELPER: &str = "conical_solid3d_contains_point_with_tolerance";
-
-        fn section<'a>(source: &'a str, start: &str, end: &str) -> &'a str {
-            let start_index = source
-                .find(start)
-                .unwrap_or_else(|| panic!("missing start marker: {start}"));
-            let tail = &source[start_index..];
-            let end_index = tail
-                .find(end)
-                .unwrap_or_else(|| panic!("missing end marker: {end}"));
-            &tail[..end_index]
-        }
-
-        let source = include_str!("primitive_3d.rs");
-        let conical_solid_point_section = section(
-            source,
-            "fn conical_solid3d_point3d_intersection_raw",
-            "fn conical_surface3d_intersect_params",
-        );
-
-        assert!(
-            conical_solid_point_section.contains(CONICAL_SOLID_POINT_HELPER),
-            "intersection/primitive_3d.rs should route conical solid point-like checks through the point helper"
-        );
-        assert!(
-            !conical_solid_point_section.contains(CONICAL_SOLID_DIRECT_CONTAINMENT),
-            "intersection/primitive_3d.rs must not call ConicalSolid3DContainment::contains_point_tolerance directly in the representative conical solid section"
-        );
-    }
-}
+mod tests;
